Handle nav history and status
The current and previously used navigation targets are implemented as indices to the waypoint/route lists. After replacing these lists with new data, those indices are no longer usable and may put the radio in an inconsistent state.

This change works around this issue by clearing the navigation status and history for data read from GPX.