Return tuple from read_atis_enabled() and read_region()
Config users should have an easy way to retrieve the ATIS enabled flag as boolean and the region as string. How these values are stored in the device differs by model, so the previous approach of referring to a dict in the memory module may not scale well when support for more devices is added.

To avoid the need for extra methods, this change lets the new (unreleased) read_* methods return both the Python-y boolean/string value as well as the actual data byte read from the device memory as a tuple. This approach is similar to that of other methods like read_mmsi().