Fix flash ID check
Since the canonical source of the flash ID is now FLASH_ID in the config classes, that's what check_flash_id() should use. With this change, the "fixme" flash_id values in the device classes can be removed.

Verifying the flash ID check using the simulator requires a way to let tests control the config memory. Automatically populating the simulator with the correct flash ID should only happen if no config data is provided to the constructor.